Refactor Render Target barrier handling for DefaultRenderer
The state management of Render Target has been moved outside
the DefaultRenderer. The application should make sure that Render
Targets passed to a renderer are in the correct state.

This was done to aviod unnecessary barreirs and to allow compound
effects of multiple renderers on a single Render Target, such as in the
case of rendering the scene and then the ImGui interface.