Refactor the codebase
The refactor got rid of old, unsused code, which mostly had to
do with elements of the old project, that will not be used here.

It also removed some code duplication from the Mesh class.
Additionally a bug was fixed, where the exit code of the application
was not acually taken from the WM_QUIT message, but rather from the message
that happened to be last in the queue.
The SceneObject class was improved by including a scale of the object
and adding some default initialization of position, scale and rotation
to its constructor.